requirements.txt
pytorch
torchvision
numpy
tensorboardX
tqdm